+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 # Change Log
 
-## [0.1.1] 2026-03-17
+## [0.1.2] 2026-03-10
 
-## Fixed
+### Changed
+
+- Added reference to `num-primitive`.
+
+- More methods in the `forward!` macro.
+
+- Added a few missing tests.
+
+## [0.1.1] 2026-03-07
+
+### Fixed
 
 - Fixed README links.
 
-## [0.1.0] 2026-03-17
+## [0.1.0] 2026-03-07
 
-## New
+### New
 
 - First release.